reservation/src/main/java/com/borealfeast/reservation/service/validation/CommonDateValidator.java
package com.borealfeast.reservation.service.validation;

import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

import java.time.LocalDate;

@Service
public class CommonDateValidator {

    public void validateDateOrder(LocalDate startDate, LocalDate endDate) {
        if (startDate.isAfter(endDate)) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid: End date is earlier than start date.");
        }
    }
}
